requirements.txt
# Sweet James Dashboard - Full Stack Requirements
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0

# Google Ads API
google-ads==28.0.0.post1

# Salesforce/Litify API
simple-salesforce==1.12.4

msal==1.31.0